+++ src/main/java/com/ixayda/iam/client/ClientAuthenticationMethod.java
@@ -1,0 +1,9 @@
+package com.ixayda.iam.client;
+
+public enum ClientAuthenticationMethod {
+
+	NONE,
+
+	CLIENT_SECRET_BASIC
+
+}