+++ src/main/java/com/ixayda/iam/client/ClientStatus.java
@@ -1,0 +1,9 @@
+package com.ixayda.iam.client;
+
+public enum ClientStatus {
+
+	ACTIVE,
+
+	DISABLED
+
+}