+++ src/main/java/com/ixayda/iam/client/ClientType.java
@@ -1,0 +1,9 @@
+package com.ixayda.iam.client;
+
+public enum ClientType {
+
+	PUBLIC,
+
+	CONFIDENTIAL
+
+}